Use double

+++ components/utils/test/irrigation/Fakes.hpp
@@ -28,16 +28,16 @@ struct FakeValve {
 };
 
 struct FakeFlow {
-    Liters bucket { 0.f };
+    Liters bucket { 0.0 };
     Liters read_and_reset_liters() {
         auto r = bucket;
-        bucket = 0.f;
+        bucket = 0.0;
         return r;
     }
 };
 
 struct FakeMoisture {
-    Percent m { 50.f };
+    Percent m { 50.0 };
     Percent read_percent() {
         return m;
     }
@@ -46,10 +46,10 @@ struct FakeMoisture {
 // Simple FOPDT-ish soil simulator (test-only)
 struct SoilSim {
     // Parameters
-    float K_percent_per_liter = 0.25f;         // % / L
+    double K_percent_per_liter = 0.25;         // % / L
     ms Td { std::chrono::seconds { 10 } };     // transport delay
     ms tau { std::chrono::seconds { 20 } };    // time constant
-    float evap_percent_per_min = 0.03f;        // natural decay when valve off
+    double evap_percent_per_min = 0.03;        // natural decay when valve off
 
     struct Input {
         ms t;
@@ -58,7 +58,7 @@ struct SoilSim {
     std::deque<Input> hist;
 
     void inject(ms now, Liters liters) {
-        if (liters > 0.f)
+        if (liters > 0.0)
             hist.push_back({ now, liters });
         // Trim very old inputs
         while (!hist.empty() && (now - hist.front().t) > (Td + 10 * tau)) {
@@ -69,24 +69,24 @@ struct SoilSim {
     // Advance one tick
     void step(ms now, Percent& m, ms dt) {
         // Evaporative drift (approximate: linear decay)
-        const float dt_min = static_cast<float>(dt.count()) / 60000.0f;
-        m = std::max(0.f, m - evap_percent_per_min * dt_min);
+        const double dt_min = static_cast<double>(dt.count()) / 60000.0;
+        m = std::max(0.0, m - evap_percent_per_min * dt_min);
 
         // Aggregate rise from all past inputs after dead-time
-        float dm_total = 0.f;
+        double dm_total = 0.0;
         for (auto const& u : hist) {
             if (now <= u.t + Td)
                 continue;    // not arrived yet
-            const float age_ms = static_cast<float>((now - (u.t + Td)).count());
-            const float tau_ms = static_cast<float>(tau.count());
-            const float rise = 1.f - std::exp(-age_ms / tau_ms);    // 0..1
-            dm_total += K_percent_per_liter * u.liters * rise;      // % contribution
+            const double age_ms = static_cast<double>((now - (u.t + Td)).count());
+            const double tau_ms = static_cast<double>(tau.count());
+            const double rise = 1.0 - std::exp(-age_ms / tau_ms);    // 0..1
+            dm_total += K_percent_per_liter * u.liters * rise;       // % contribution
         }
 
         // Crude discrete application so moisture approaches the simulated target smoothly
         // Scale by dt/tau (bounded) to avoid big jumps on large dt.
-        const float scale = std::min(1.0f, static_cast<float>(dt.count()) / static_cast<float>(tau.count()));
-        m = std::min(100.f, m + dm_total * 0.1f * scale);
+        const double scale = std::min(1.0, static_cast<double>(dt.count()) / static_cast<double>(tau.count()));
+        m = std::min(100.0, m + dm_total * 0.1 * scale);
     }
 };
